+++ interface/index.html
@@ -179,6 +179,6 @@
         </div>
       </div>
     </main>
-    <script type="module" src="./src/main.js"></script>
+    <script type="module" src="./src/Simulador/main.js"></script>
   </body>
 </html>
