+++ receptor/src/CamadaEnlace/checkBitCounting.ts
