+++ receptor/src/CamadaEnlace/checkErrorControl.ts
