+++ receptor/src/CamadaFisica/decode.ts
