+++ receptor/src/CamadaFisica/textFromBits.ts
