+++ transmitter/src/CamadaEnlace/bitCounting.ts
