+++ transmitter/src/CamadaEnlace/errorControl.ts
