+++ transmitter/src/CamadaFisica/bitsFromText.ts
