+++ transmitter/src/CamadaFisica/encode.ts
