+++ transmitter/src/InterfaceGUI/interfaceComms.ts
@@ -1,5 +1,5 @@
 import type { Response, Request, Query } from "express-serve-static-core";
-import type { EncodingType } from "./Física/encode";
+import type { EncodingType } from "../CamadaFisica/encode";
 
 type Res = Response<any, Record<string, any>, number>;
 type Req = Request<{}, any, any, Query, Record<string, any>>;
